登陆增加角色；Jwt增加Redis存储；

+++ lveqia-cloud-zuul/src/main/java/com/lveqia/cloud/zuul/mapper/SysMenuMapper.java
@@ -60,5 +60,5 @@ public interface SysMenuMapper {
     @ResultMap("sysMenu")
     @Select("SELECT * FROM t_sys_menu WHERE `id` IN(SELECT mr.`mid` FROM t_sys_user_role ur,t_sys_menu_role mr " +
             " WHERE ur.`rid`=mr.`rid` AND ur.`uid` =#{id}) AND `enabled`=true")
-    List<SysMenu> getMenusByUserId(@Param("id") Integer id);
+    List<SysMenu> getMenusByUserId(@Param("id") Long id);
 }
